Query MongoDB through collections, not raw commands (#3)
* refactor(data): query MongoDB through collections, not raw commands

Every question this application asked was a command document it assembled
itself: `aggregate` with its cursor options, `insert` with `ordered`,
`createIndexes` with both specifications, `drop`, and a `$count`
pipeline read back by hand. The library now carries all of that, so the
data layer keeps the part that is about coffee places and hands the rest
over.

`MongoSeam` is gone. `:data` depends on `embedded-mongodb-core` -- the
library's Android-free half -- and a `PlaceRepository` is built from a
`MongoCollection`, a `Seeder` from two. The command builders become
pipeline builders: `nearestPipeline`, `searchPipeline`,
`viewportPipeline` and `placeIndexes` are what this application knows and
nothing more. Seeding inserts with `insertMany(ordered = false)`, counts
with `countDocuments`, and writes its marker with `insertOne`.

The tests move down with them. `FakeMongo` is now a `CommandRunner`, so
what a test scripts is the seam the library itself is written against
rather than one this application invented, and the assertions are on the
commands the library builds around a pipeline.

`SeedProgress` counts are `Long`, because `countDocuments` reports one
and narrowing it would be a truncation nobody asked for.

`:data` still needs no emulator and no compiled engine; it does now need
the library checkout, because the core module is where the collections
live.


* docs: describe the data layer's move onto the collection API

The layout, the seam section and the substitution block all described a
`MongoSeam` that no longer exists. The seam section becomes an account of
what this application still builds -- the pipelines, the index
specifications and the parsing -- and what the library now builds around
it.

The hardware measurements are marked as predating the change: the
commands reaching the engine are the same documents, and the pipeline
tests pin them, but the figures have not been re-taken on the phone.


* refactor(data): read documents through the query rather than through map

`PlaceRepository` collected `asFlow().map(read).toList()`; the library now
takes the reading function directly, so it is `asFlow(read).toList()` and
the parsing sits next to the query it belongs to.

`Seeder` imports `drop`, which moved out of `MongoCollection` and into an
extension alongside every other operation.


* refactor(data): follow the library onto the drivers' names

`getDatabase`/`getCollection`, and the text index becomes
`compoundIndex(text("name"), text("brand"))` -- a collection may hold
only one text index, and that is how the driver spells one over two
fields. The index built is the same index; `PlaceIndexesTest` pins the
command field for field.


---------

+++ data/src/main/kotlin/io/github/jeroenvervaeke/coffeefinder/data/seed/SeedProgress.kt
@@ -11,11 +11,11 @@ sealed interface SeedProgress {
     data object Checking : SeedProgress
 
     /** [inserted] documents have reached the collection so far. */
-    data class Inserting(val inserted: Int) : SeedProgress
+    data class Inserting(val inserted: Long) : SeedProgress
 
     /** Building the `2dsphere` and text indexes, which happens after the documents are in. */
     data object Indexing : SeedProgress
 
     /** The database holds [places] coffee places and is ready to be queried. */
-    data class Ready(val places: Int) : SeedProgress
+    data class Ready(val places: Long) : SeedProgress
 }
